Drew - product bean 裡面的 productid 從 Integer 改成 Long。 有用到的記得改! 需要幫忙改的也可以跟我說! 我幫你改

+++ src/main/java/foodelicious/compbackend/model/CBKProductDaoInterface.java
@@ -11,11 +11,11 @@ public interface CBKProductDaoInterface {
 
 	List<Product> getAllProducts(HttpSession session);
 
-	String deleteProduct(Integer productId);
+	String deleteProduct(Long productId);
 
-	Product findByProductId(Integer productId);
+	Product findByProductId(Long productId);
 	
-	String updateProduct(Integer productId, Product product);
+	String updateProduct(Long productId, Product product);
